Added messaging, removed API call
Added a dialog display if the app was not found when the script ran
(the app name may simply not match or there could be another problem).

Removed API call and code to lookup the user’s JSS ID. This is
unneeded.  The PUT operation can be performed on a ‘/name/‘ endpoint.

Removed import of ‘xml.etree.cElementTree’ as it was no longer needed
for parsing any return XML.